feat: add AnimatableHeaderShadow component
