+++ src/lib/Components/AnimatableHeader/AnimatableHeaderFlatList.tsx
@@ -3,6 +3,7 @@ import { FlatList, FlatListProps } from "react-native"
 import Animated from "react-native-reanimated"
 import { useAnimatableHeaderContext } from "./AnimatableHeaderContext"
 import { AnimatableHeaderLargeTitle } from "./AnimatableHeaderLargeTitle"
+import { AnimatableHeaderShadow } from "./AnimatableHeaderShadow"
 
 const AnimatedComponent: typeof FlatList = Animated.createAnimatedComponent(FlatList)
 
@@ -11,22 +12,25 @@ export const AnimatableHeaderFlatList = <T extends any>(props: FlatListProps<T>)
   const { scrollOffsetY } = useAnimatableHeaderContext()
 
   return (
-    <AnimatedComponent
-      {...other}
-      scrollEventThrottle={0.0000000001}
-      ListHeaderComponent={
-        <>
-          <AnimatableHeaderLargeTitle />
-          {ListHeaderComponent}
-        </>
-      }
-      onScroll={Animated.event([
-        {
-          nativeEvent: {
-            contentOffset: { y: scrollOffsetY },
+    <>
+      <AnimatedComponent
+        {...other}
+        scrollEventThrottle={0.0000000001}
+        ListHeaderComponent={
+          <>
+            <AnimatableHeaderLargeTitle />
+            {ListHeaderComponent}
+          </>
+        }
+        onScroll={Animated.event([
+          {
+            nativeEvent: {
+              contentOffset: { y: scrollOffsetY },
+            },
           },
-        },
-      ])}
-    />
+        ])}
+      />
+      <AnimatableHeaderShadow />
+    </>
   )
 }
