+++ src/lib/Components/AnimatableHeader/AnimatableHeaderScrollView.tsx
@@ -3,6 +3,7 @@ import { ScrollView, ScrollViewProps } from "react-native"
 import Animated from "react-native-reanimated"
 import { useAnimatableHeaderContext } from "./AnimatableHeaderContext"
 import { AnimatableHeaderLargeTitle } from "./AnimatableHeaderLargeTitle"
+import { AnimatableHeaderShadow } from "./AnimatableHeaderShadow"
 
 const AnimatedComponent: typeof ScrollView = Animated.createAnimatedComponent(ScrollView)
 
@@ -11,19 +12,22 @@ export const AnimatableHeaderScrollView: React.FC<ScrollViewProps> = (props) => 
   const { scrollOffsetY } = useAnimatableHeaderContext()
 
   return (
-    <AnimatedComponent
-      {...other}
-      scrollEventThrottle={0.0000000001}
-      onScroll={Animated.event([
-        {
-          nativeEvent: {
-            contentOffset: { y: scrollOffsetY },
+    <>
+      <AnimatedComponent
+        {...other}
+        scrollEventThrottle={0.0000000001}
+        onScroll={Animated.event([
+          {
+            nativeEvent: {
+              contentOffset: { y: scrollOffsetY },
+            },
           },
-        },
-      ])}
-    >
-      <AnimatableHeaderLargeTitle />
-      {children}
-    </AnimatedComponent>
+        ])}
+      >
+        <AnimatableHeaderLargeTitle />
+        {children}
+      </AnimatedComponent>
+      <AnimatableHeaderShadow />
+    </>
   )
 }
